refactor: splitt characterEquipment.ts i domenefiler (#94)
732-linjers monolitt splittet etter routes/scenarios/-mønsteret:
- parse.ts      — alle parseXxx-funksjoner
- access.ts     — requireAccessibleCharacter
- helpers.ts    — toEquipmentFeatureState, loadDevSampleCharacterEquipment
- itemRoutes.ts — GET state + item-ruter (move, add, remove, quantity, metadata)
- locationRoutes.ts — lokasjonsstyring (opprett, slett)
- loadoutRoutes.ts  — loadout-ruter (rustning, skjold, våpenslots)
- devRoutes.ts  — bootstrap-sample (ikke-produksjon)
- routes.ts     — hoved-plugin som registrerer sub-plugins
- index.ts      — re-eksport av characterEquipmentRoutes

Ingen logikkendringer. app.ts-importen er uendret.
