+++ apps/api/src/routes/characterEquipment/index.ts
@@ -1,0 +1,1 @@
+export { characterEquipmentRoutes } from "./routes";
